어트리뷰트 타입 정수에서 실수로 수정
1. 스킬 닌자머리 닌자바디 구현
2. 어트리뷰트 타입 정수에서 실수로 수정
 - 이유는 스킬 확장성에서 스탯이 정수로는 어려움을 느낌
예시 공격속도를 30% 증가시킨다 기존 어트리뷰트에서는 정수라 30%가 불가능

+++ Assets/Scripts/LocalizationManager/LocalizedText.cs
@@ -47,7 +47,7 @@ public class LocalizedText : MonoBehaviour
         textComponent.text = localizedText;
     }
 
-    public void UpdateLocalizedText(int value)
+    public void UpdateLocalizedText(float value)
     {
         if (string.IsNullOrEmpty(localizationID))
         {
@@ -63,7 +63,7 @@ public class LocalizedText : MonoBehaviour
         textComponent.text = localizedText;
     }
 
-    public void UpdateLocalizedText(int value1, int value2)
+    public void UpdateLocalizedText(float value1, float value2)
     {
         if (string.IsNullOrEmpty(localizationID))
         {
